Markdown instead of scratch
I changed the Magic file, and now we are using Markdown (.md) files
Instead of my hand made Content (.con) file.

Surly It is better cause if i wanted to make a whole another Markdown
Like format it would be a waste of time.

There is a special 2 lines in each content files that related to `title`
And `Info`